+++ bin/capture.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-# Copyright 2020-2021 Jetperch LLC
+# Copyright 2020-2025 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,103 +15,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Capture Joulescope data to a JLS v2 file.  See https://github.com/jetperch/jls"""
 
-from joulescope import scan, scan_require_one, JlsWriter
-from joulescope.data_recorder import DataRecorder
-from joulescope.units import duration_to_seconds
+"""Capture Joulescope data to a JLS v2 file.
+
+Instead of this script, consider using pyjoulescope_driver directly, like this:
+
+    python -m pyjoulescope_driver capture out.jls
+
+See https://github.com/jetperch/jls and
+https://github.com/jetperch/joulescope_driver/blob/main/pyjoulescope_driver/entry_points/record.py
+"""
+
+
+from pyjoulescope_driver.entry_points import record
 import argparse
-import signal
-import time
-
-
-SIGNALS = {
-    'current': (1, 'A'),
-    'voltage': (2, 'V'),
-    'power': (3, 'W'),
-}
+import sys
 
 
 def get_parser():
-    p = argparse.ArgumentParser(
-        description='Capture Joulescope samples to a JLS file.')
-    p.add_argument('--duration',
-                   type=duration_to_seconds,
-                   help='The capture duration in float seconds. '
-                   + 'Add a suffix for other units: s=seconds, m=minutes, h=hours, d=days')
-    p.add_argument('--frequency', '-f',
-                   help='The sampling frequency in Hz.')
-    p.add_argument('--jls',
-                   default=1,
-                   type=int,
-                   choices=[1, 2],
-                   help='The JLS file format version.  For v2, see https://github.com/jetperch/jls')
-    p.add_argument('--serial_number',
-                   help='The serial number of the Joulescope for this capture.')
-    p.add_argument('--signals',
-                   default='current,voltage',
-                   help='The comma-separated list of signals to capture which include current, voltage, power. '
-                   + 'Defaults to current,voltage.  This setting only applies to jls v2.')
-    p.add_argument('filename',
-                   help='The JLS filename to record.')
-    return p
-
-
-def scan_by_serial_number(serial_number, name: str = None, config=None):
-    devices = scan(name, config)
-    for device in devices:
-        if serial_number == device.device_serial_number:
-            return device
-    raise KeyError(f'Device not found with serial number {serial_number}')
-
-
-def run():
-    quit_ = False
-    args = get_parser().parse_args()
-    duration = args.duration
-
-    def do_quit(*args, **kwargs):
-        nonlocal quit_
-        quit_ = 'quit from SIGINT'
-
-    signal.signal(signal.SIGINT, do_quit)
-    
-    if args.serial_number is not None:
-        device = scan_by_serial_number(args.serial_number, config='auto')
-    else:
-        device = scan_require_one(config='auto')
-
-    if args.frequency:
-        try:
-            device.parameter_set('sampling_frequency', int(args.frequency))
-        except Exception:
-            # bad frequency selected, display warning & exit gracefully
-            freqs = [f[2][0] for f in device.parameters('sampling_frequency').options]
-            print(f'Unsupported frequency selected: {args.frequency}', flush=True)
-            print(f'Supported frequencies = {freqs}', flush=True)
-            return 1
-
-    with device:
-        if args.jls == 1:
-            recorder = DataRecorder(args.filename, calibration=device.calibration)
-        else:
-            recorder = JlsWriter(device, args.filename, signals=args.signals)
-            recorder.open()
-        try:
-            device.stream_process_register(recorder)
-            t_stop = None if duration is None else time.time() + duration
-            device.start()
-            print(f'Capturing data from {device}: type CTRL-C to stop', flush=True)
-            while not quit_:
-                time.sleep(0.01)
-                if t_stop and time.time() > t_stop:
-                    break
-            device.stop()
-        finally:
-            recorder.close()
-            device.stream_process_unregister(recorder)
-    return 0
+    parser = argparse.ArgumentParser(
+        description='Capture streaming samples to a JLS v2 file.',
+    )
+    parser.add_argument('--jsdrv_log_level',
+                        choices=['off', 'emergency', 'alert', 'critical', 'error', 'warning',
+                                 'notice', 'info', 'debug1', 'debug2', 'debug3', 'all'],
+                        default='error',
+                        help='Configure the joulescope driver native log level.')
+    record.parser_config(parser)
+    return parser
 
 
 if __name__ == '__main__':
-    run()
+    args = get_parser().parse_args()
+    sys.exit(record.on_cmd(args))