+++ CHANGELOG.md
@@ -1,7 +1,32 @@
+<!--
+# Copyright 2019-2025 Jetperch LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+-->
+
 
 # CHANGELOG
 
 This file contains the list of changes made to pyjoulescope_examples.
+
+
+## 0.14.0
+
+2025 Jan 2
+
+* Added bin/fuse_js220.py.
+* Added bin/timesync.py
+* Updated capture.py to use pyjoulescope_driver record entry point.
 
 
 ## 0.13.0